# src/main/java/org/Autogenerated/src/main/java/io/codejournal/maven/swagger2java/api/BookToBookStoreApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.28).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.codejournal.maven.swagger2java.api;

import io.codejournal.maven.swagger2java.model.BookToBookStore;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-09-09T15:04:09.579+05:30")

@Validated
@Api(value = "BookToBookStore", description = "the BookToBookStore API")
@RequestMapping(value = "/v2")
public interface BookToBookStoreApi {

    @ApiOperation(value = "Add a new BookToBookStore object", nickname = "addBookToBookStore", notes = "Add a new BookToBookStore object value.", tags={ "BookToBookStore", })
    @ApiResponses(value = { 
        @ApiResponse(code = 405, message = "Invalid input") })
    @RequestMapping(value = "/BookToBookStore",
        produces = { "application/xml", "application/json" }, 
        consumes = { "application/json", "application/xml" },
        method = RequestMethod.POST)
    ResponseEntity<Void> addBookToBookStore(@ApiParam(value = "Pet object that needs to be added to the store" ,required=true )  @Valid @RequestBody BookToBookStore body);


    @ApiOperation(value = "Deletes a BookToBookStore", nickname = "deleteBookToBookStore", notes = "", tags={ "BookToBookStore", })
    @ApiResponses(value = { 
        @ApiResponse(code = 400, message = "Invalid BookToBookStore value") })
    @RequestMapping(value = "/BookToBookStore/{NAME}",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteBookToBookStore(@ApiParam(value = "Pet BookToBookStore to delete",required=true) @PathVariable("NAME") String NAME);


    @ApiOperation(value = "Find BookToBookStore by NAME", nickname = "getBookToBookStoreByNAME", notes = "Returns a single BookToBookStore", response = BookToBookStore.class, tags={ "BookToBookStore", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = BookToBookStore.class),
        @ApiResponse(code = 400, message = "Invalid NAME supplied"),
        @ApiResponse(code = 404, message = "BookToBookStore not found") })
    @RequestMapping(value = "/BookToBookStore/{NAME}",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<BookToBookStore> getBookToBookStoreByNAME(@ApiParam(value = "ID of BookToBookStore to return",required=true) @PathVariable("NAME") String NAME);

}