# src/main/java/org/Autogenerated/src/main/java/io/codejournal/maven/swagger2java/api/BookToBookStoreApiController.java
package io.codejournal.maven.swagger2java.api;

import io.codejournal.maven.swagger2java.model.BookToBookStore;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.*;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.constraints.*;
import javax.validation.Valid;
import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-09-09T15:04:09.579+05:30")

@Controller
public class BookToBookStoreApiController implements BookToBookStoreApi {

    private static final Logger log = LoggerFactory.getLogger(BookToBookStoreApiController.class);

    private final ObjectMapper objectMapper;

    private final HttpServletRequest request;

    @org.springframework.beans.factory.annotation.Autowired
    public BookToBookStoreApiController(ObjectMapper objectMapper, HttpServletRequest request) {
        this.objectMapper = objectMapper;
        this.request = request;
    }

    public ResponseEntity<Void> addBookToBookStore(@ApiParam(value = "Pet object that needs to be added to the store" ,required=true )  @Valid @RequestBody BookToBookStore body) {
        String accept = request.getHeader("Accept");
        return new ResponseEntity<Void>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<Void> deleteBookToBookStore(@ApiParam(value = "Pet BookToBookStore to delete",required=true) @PathVariable("NAME") String NAME) {
        String accept = request.getHeader("Accept");
        return new ResponseEntity<Void>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<BookToBookStore> getBookToBookStoreByNAME(@ApiParam(value = "ID of BookToBookStore to return",required=true) @PathVariable("NAME") String NAME) {
        String accept = request.getHeader("Accept");
        if (accept != null && accept.contains("application/xml")) {
            try {
                return new ResponseEntity<BookToBookStore>(objectMapper.readValue("<BookToBookStore>  <NAME>aeiou</NAME>  <BOOK_ID>123456789</BOOK_ID>  <STOCK>123456789</STOCK></BookToBookStore>", BookToBookStore.class), HttpStatus.NOT_IMPLEMENTED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/xml", e);
                return new ResponseEntity<BookToBookStore>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        if (accept != null && accept.contains("application/json")) {
            try {
                return new ResponseEntity<BookToBookStore>(objectMapper.readValue("{\"empty\": false}", BookToBookStore.class), HttpStatus.NOT_IMPLEMENTED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/json", e);
                return new ResponseEntity<BookToBookStore>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        return new ResponseEntity<BookToBookStore>(HttpStatus.NOT_IMPLEMENTED);
    }

}