# src/main/java/org/Autogenerated/src/main/java/io/codejournal/maven/swagger2java/api/LanguageApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.28).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.codejournal.maven.swagger2java.api;

import io.codejournal.maven.swagger2java.model.Language;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-09-09T15:04:09.579+05:30")

@Validated
@Api(value = "Language", description = "the Language API")
@RequestMapping(value = "/v2")
public interface LanguageApi {

    @ApiOperation(value = "Add a new Language object", nickname = "addLanguage", notes = "Add a new Language object value.", tags={ "Language", })
    @ApiResponses(value = { 
        @ApiResponse(code = 405, message = "Invalid input") })
    @RequestMapping(value = "/Language",
        produces = { "application/xml", "application/json" }, 
        consumes = { "application/json", "application/xml" },
        method = RequestMethod.POST)
    ResponseEntity<Void> addLanguage(@ApiParam(value = "Pet object that needs to be added to the store" ,required=true )  @Valid @RequestBody Language body);


    @ApiOperation(value = "Deletes a Language", nickname = "deleteLanguage", notes = "", tags={ "Language", })
    @ApiResponses(value = { 
        @ApiResponse(code = 400, message = "Invalid Language value") })
    @RequestMapping(value = "/Language/{ID}",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteLanguage(@ApiParam(value = "Pet Language to delete",required=true) @PathVariable("ID") Long ID);


    @ApiOperation(value = "Find Language by ID", nickname = "getLanguageByID", notes = "Returns a single Language", response = Language.class, tags={ "Language", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Language.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Language not found") })
    @RequestMapping(value = "/Language/{ID}",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Language> getLanguageByID(@ApiParam(value = "ID of Language to return",required=true) @PathVariable("ID") Long ID);

}